refactor: Structure the unstructured chaos
