+++ lib/screens/settings_screen/widgets/timer_mode_toggle.dart
