+++ lib/screens/settings_screen/widgets/toggle_hide_settings_button.dart
