+++ lib/screens/timer_screen/orodomop/timer_control_row.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:orodomop/providers/timer_provider.dart';
 import 'package:orodomop/models/timer_state.dart';
-import 'package:orodomop/widgets/buttons/reset_button.dart';
-import 'package:orodomop/widgets/pomodoro/break_button.dart';
-import 'package:orodomop/widgets/buttons/end_break_button.dart';
-import 'package:orodomop/widgets/buttons/pause_or_resume_button.dart';
-import 'package:orodomop/widgets/buttons/start_button.dart';
+import 'package:orodomop/screens/timer_screen/orodomop/break_button.dart';
+import 'package:orodomop/screens/timer_screen/widgets/buttons/end_break_button.dart';
+import 'package:orodomop/screens/timer_screen/widgets/buttons/pause_or_resume_button.dart';
+import 'package:orodomop/screens/timer_screen/widgets/buttons/reset_button.dart';
+import 'package:orodomop/screens/timer_screen/widgets/buttons/start_button.dart';
 import 'package:provider/provider.dart';
 
 class TimerControlRow extends StatelessWidget {
@@ -24,15 +24,8 @@ class TimerControlRow extends StatelessWidget {
             timerState: timerModel.timerState,
           ),
       builder: (context, values, child) {
-        bool onBreak =
-            (values.breakTime > 0 && values.timerState.isPaused) ||
-            values.timerState.isOnBreak;
-
-        bool onFocus =
-            (values.focusTime > 0 && values.timerState.isPaused) ||
-            values.timerState.isOnFocus;
-
-        if (values.timerState.isIdle && values.focusTime > 0) {
+        if (values.timerState.isIdle ||
+            values.timerState.isOnBreak && values.breakTime == 0) {
           return StartButton();
         }
 
@@ -41,11 +34,17 @@ class TimerControlRow extends StatelessWidget {
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                onBreak || onFocus ? PauseOrResumeButton() : BreakButton(),
+                PauseOrResumeButton(),
                 SizedBox(width: 16),
-                onFocus ? ResetButton() : EndBreakButton(),
+                (values.breakTime > 0 && values.timerState.isPaused) ||
+                        values.timerState.isOnBreak
+                    ? EndBreakButton()
+                    : BreakButton(),
               ],
             ),
+            values.timerState.isOnFocus || values.focusTime > 0
+                ? ResetButton()
+                : SizedBox.shrink(),
           ],
         );
       },
