+++ lib/screens/timer_screen/widgets/buttons/reset_button.dart
